// src/app/interfaces/ConfigurationData.ts
export interface ConfigurationData {
    advises: string
    contact_name: string
    customer_id: number
    delivery_conditions: number
    id: number
    minimal_distance: string
    observations: string
    peripheries: string
    rate_base: string
    rate_per_minute: string
    schedule: string
    time_wait: string
    value_per_distance: string
    transport_type_id : number
    km_base : string
    neighboring_base : string
    km_unit_value : string

}